Refactor Hilt DI configuration across core and app modules (PR #537 comments)
- Delete data:common and move CommonModule and AppQualifiers to app module
- Reorganize low-light, low-light-playservices, and postprocess DI modules into core:camera:*:*-di sub-modules
- Revert low-light-playservices to core and remove Hilt/kapt plugins from base capability modules
- Remove @Inject and @Qualifier annotations from core constructors and instantiate core classes via @Provides
- Update settings.gradle.kts and app module dependencies
