+++ app/build.gradle.kts
@@ -101,7 +101,11 @@ android {
 dependencies {
     implementation(libs.androidx.tracing)
     implementation(project(":core:common"))
-    implementation(project(":data:common"))
+    implementation(project(":core:camera"))
+    implementation(project(":core:camera:low-light"))
+    implementation(project(":core:camera:postprocess"))
+    implementation(project(":data:camera"))
+    implementation(project(":data:media"))
     implementation(project(":feature:postcapture"))
     // Compose
     val composeBom = platform(libs.compose.bom)
@@ -179,11 +183,11 @@ dependencies {
     implementation(project(":ui:debug"))
 
     // Low Light implementations
-    implementation(project(":data:camera:low-light"))
-    implementation(project(":data:camera:low-light-playservices"))
+    implementation(project(":core:camera:low-light:low-light-di"))
+    implementation(project(":core:camera:low-light-playservices-di"))
 
     // Postprocess implementations
-    implementation(project(":data:camera:postprocess"))
+    implementation(project(":core:camera:postprocess:postprocess-di"))
 
     implementation(project(":core:camera:low-light-playservices"))
     implementation(project(":core:camera:effects:single-stream"))
