+++ core/camera/postprocess/postprocess-di/build.gradle.kts
@@ -22,37 +22,11 @@ plugins {
 }
 
 android {
-    namespace = "com.google.jetpackcamera.data.camera"
+    namespace = "com.google.jetpackcamera.core.camera.postprocess.di"
     compileSdk = libs.versions.compileSdk.get().toInt()
 
     defaultConfig {
         minSdk = libs.versions.minSdk.get().toInt()
-        testOptions.targetSdk = libs.versions.targetSdk.get().toInt()
-        lint.targetSdk = libs.versions.targetSdk.get().toInt()
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-        create("benchmark") {
-            initWith(buildTypes.getByName("release"))
-            matchingFallbacks += listOf("release")
-        }
-    }
-
-    flavorDimensions += "flavor"
-    productFlavors {
-        create("stable") {
-            dimension = "flavor"
-            isDefault = true
-        }
     }
 
     compileOptions {
@@ -65,18 +39,13 @@ android {
 }
 
 dependencies {
+    implementation(project(":core:camera:postprocess"))
+
+    // Hilt
     implementation(libs.dagger.hilt.android)
     kapt(libs.dagger.hilt.compiler)
-    implementation(project(":core:camera"))
-    implementation(project(":core:camera:low-light"))
-    implementation(project(":core:camera:postprocess"))
-    implementation(project(":core:common"))
-    implementation(project(":core:model"))
-    implementation(project(":data:settings"))
-    implementation(project(":core:settings"))
 }
 
-// Allow references to generated code
 kapt {
     correctErrorTypes = true
 }
