--- data/camera/postprocess/src/main/java/com/google/jetpackcamera/data/camera/postprocess/PostProcessModule.kt
@@ -1,55 +1,0 @@
-/*
- * Copyright (C) 2025 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.jetpackcamera.data.camera.postprocess
-
-import com.google.jetpackcamera.core.camera.postprocess.ImagePostProcessor
-import com.google.jetpackcamera.core.camera.postprocess.ImagePostProcessorFeatureKey
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import dagger.multibindings.Multibinds
-import javax.inject.Provider
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class PostProcessModule {
-    @Multibinds
-    abstract fun imagePostProcessorEntries(): Set<
-        Map.Entry<
-            ImagePostProcessorFeatureKey,
-            @JvmSuppressWildcards Provider<ImagePostProcessor>
-            >
-        >
-
-    companion object {
-        @Provides
-        @Singleton
-        fun provideImagePostProcessorMap(
-            entries: @JvmSuppressWildcards Set<
-                Map.Entry<
-                    ImagePostProcessorFeatureKey,
-                    @JvmSuppressWildcards Provider<ImagePostProcessor>
-                    >
-                >
-        ): Map<
-            ImagePostProcessorFeatureKey,
-            @JvmSuppressWildcards Provider<ImagePostProcessor>
-            > =
-            entries.associate { it.key to it.value }
-    }
-}
