--- data/common/consumer-rules.pro
@@ -1,1 +1,0 @@
-# Proguard rules for data:common
