+++ data/media/src/main/kotlin/com/google/jetpackcamera/data/media/LocalMediaRepository.kt
@@ -28,13 +28,9 @@ import android.provider.MediaStore
 import android.util.Log
 import android.util.Size
 import androidx.core.net.toFile
-import com.google.jetpackcamera.core.common.DefaultFilePathGenerator
 import com.google.jetpackcamera.core.common.FilePathGenerator
-import com.google.jetpackcamera.core.common.IODispatcher
-import dagger.hilt.android.qualifiers.ApplicationContext
 import java.io.File
 import java.io.IOException
-import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -48,11 +44,10 @@ private const val TAG = "LocalMediaRepository"
 private const val IMAGE_MIME_TYPE = "image/jpeg"
 private const val VIDEO_MIME_TYPE = "video/mp4"
 
-class LocalMediaRepository
-@Inject constructor(
-    @ApplicationContext private val context: Context,
-    @IODispatcher private val iODispatcher: CoroutineDispatcher,
-    @DefaultFilePathGenerator private val filePathGenerator: FilePathGenerator
+class LocalMediaRepository(
+    private val context: Context,
+    private val iODispatcher: CoroutineDispatcher,
+    private val filePathGenerator: FilePathGenerator
 ) : MediaRepository {
     private val repositoryScope = CoroutineScope(iODispatcher + SupervisorJob())
     private val _currentMedia = MutableStateFlow<MediaDescriptor>(MediaDescriptor.None)
