--- data/media/src/main/kotlin/com/google/jetpackcamera/data/media/MediaModule.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright (C) 2025 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.jetpackcamera.data.media
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-/**
- * Dagger [Module] for Media dependencies.
- */
-@Module
-@InstallIn(SingletonComponent::class)
-interface MediaModule {
-
-    @Binds
-    @Singleton
-    fun bindsMediaRepository(localMediaRepository: LocalMediaRepository): MediaRepository
-}
