+++ settings.gradle.kts
@@ -39,16 +39,15 @@ include(":core:camera")
 include(":data:camera")
 include(":core:camera:testing")
 include(":core:camera:low-light")
+include(":core:camera:low-light:low-light-di")
 include(":core:camera:low-light-playservices")
+include(":core:camera:low-light-playservices-di")
 include(":core:camera:effects:single-stream")
-include(":data:camera:low-light")
-include(":data:camera:low-light-playservices")
 include(":feature:settings")
 include(":data:settings")
 include(":data:settings:testing")
 include(":data:media")
 include(":data:media:testing")
-include(":data:common")
 include(":core:common")
 include(":core:common:testing")
 include(":benchmark")
@@ -66,7 +65,7 @@ include(":core:model")
 include(":ui:uistate:postcapture")
 include(":ui:uistateadapter:postcapture")
 include(":core:camera:postprocess")
-include(":data:camera:postprocess")
+include(":core:camera:postprocess:postprocess-di")
 include(":ui:controller")
 include(":ui:controller:impl")
 include(":ui:controller:testing")
